feat: getStorefront method implementation for fetching App Store and Google Play region (#2877)
The purpose of this PR is to add the ability to determine which App Store or Play Store region user in. In my case i have different in-app purchases and subscriptions packs for different regions. And in order to offer users specific purchases it's necessary to know their Store region.

In addition Android getBillingConfigAsync() method should be used with existing BillingClient, so, in my opinion, the implementation of this feature suggests itself. Thanks in advance!

https://developer.android.com/reference/com/android/billingclient/api/BillingClient#getBillingConfigAsync(com.android.billingclient.api.GetBillingConfigParams,com.android.billingclient.api.BillingConfigResponseListener)